bulk.c (tds_bcp_send_record): Reject invalid NULL data.
When the user erroneously supplies NULL data for a column which does not
use a nullable type and is not otherwise known to be nullable, cleanly
fail.  Pressing on anyway (in hopes that a server-side default would
take effect?) was liable to trigger legitimate assertion failures in
tds_checks.c.